+++ src/command/command.ts
@@ -1,6 +1,7 @@
 import { ExtensionContext, commands } from "vscode";
 
 import { setApiToken, setShowOfflineRobot } from "./globalConfig";
+import { uploadScript } from "./operation";
 import {
     setProjectName,
     markConfigFile,
@@ -21,5 +22,6 @@ export function registerCommand(context: ExtensionContext): void {
         commands.registerCommand("hamibot-assistant.setExecuteRobot", setExecuteRobot),
 
         // 操作
+        commands.registerCommand("hamibot-assistant.uploadScript", uploadScript)
     );
 }