+++ CHANGELOG.md
@@ -10,8 +10,14 @@
 ### Added
 
 - 全局设置指令:
+
   - `hamibot-assistant.setApiToken` : 设置用于访问 Hamibot API 的开发者令牌
   - `hamibot-assistant.setShowOfflineRobot` : 设置是否显示离线机器人
+
+- 项目操作指令：
+
+  - `hamibot-assistant.uploadFile` : 上传标记的脚本文件
+
 - 更详细的请求失败报错
 - 访问接口前检查请求参数
 - 上传文件接口
@@ -32,11 +38,15 @@
 - 项目设置 Schema ( `hamibot.config.json` )
 - 项目设置操作接口
 - 项目设置指令:
+
   - `hamibot-assistant.setProjectName` : 设置项目名称
   - `hamibot-assistant.setExecuteRobot` : 设置调试机器人
+
 - 文件树右键菜单快捷方式:
+
   - 标记脚本文件（ `.js` ）
   - 标记配置文件（ `.json` ）
+
 - 除上传文件以外全部 Hamibot 接口的封装
 - 服务器错误时尝试使用备用的域名请求